==> 02_NSW_Charger_Optimizer/step1_merge_traffic_info.ipynb <==
# 这个文件用来统计车流量
###### 从01_data_for_all_year.csv中提取traffic_count记录在01_station_information.csv中，记作新csv文件
###### 同时从01_Copy_of_Heavy_traffic_counter.csv中检查每个站点是否包含重型交通

import pandas as pd

# 读取第一个CSV文件
df1 = pd.read_csv('dataset/01_Traffic_Volume_Viewer_-_Data_for_All_Years.csv')

# 读取第二个CSV文件
df2 = pd.read_csv('dataset/01_Traffic_Volume_Viewer_-_Station_Information.csv')

df3 = pd.read_csv('dataset/01_Copy_of_Heavy_traffic_counter.csv')

# df1.head()

# filtered_df = df1[(df1['cardinal_direction_name'].isin(['BOTH', 'EASTBOUND AND WESTBOUND', 'NORTHBOUND AND SOUTHBOUND'])) & (df1['period'] == 'ALL DAYS')]
max_year_per_station = df1.groupby('station_id')['year'].max().reset_index()
all_days_max_year = pd.merge(df1, max_year_per_station, on=['station_id', 'year'])
all_days_max_year = all_days_max_year[all_days_max_year['period'] == 'ALL DAYS']


# filtered_df.head(10)

# max_year_per_station = filtered_df.groupby('station_id')['year'].max().reset_index()
# result = pd.merge(filtered_df, max_year_per_station, on=['station_id', 'year'])
filtered_df = df1[(df1['cardinal_direction_name'].isin(['BOTH', 'EASTBOUND AND WESTBOUND', 'NORTHBOUND AND SOUTHBOUND']))
                  & (df1['period'] == 'ALL DAYS')]
filtered_max_year = pd.merge(filtered_df, max_year_per_station, on=['station_id', 'year'])

result = pd.concat([all_days_max_year, filtered_max_year]).drop_duplicates()


result.head(20)

# 这时station_id仍有重复出现的情况，当station_id重复出现时，应只保留“traffic_count”最大的那条数据
result_sorted = result.sort_values(by=['station_id', 'traffic_count'], ascending=[True, False])

final_result = result_sorted.groupby('station_id').head(1)


final_result

merged_df = pd.merge(df2, final_result[['station_id', 'traffic_count']], on='station_id', how='left')
merged_df['traffic_count'] = merged_df['traffic_count'].fillna(0)        # 不确定是否有缺失值，加上保险


df4 = merged_df[['station_id', 'full_name', 'suburb', 'post_code', 'wgs84_latitude', 'wgs84_longitude', 'traffic_count']]


df4

df4.info()

# df4.describe()

df4['station_id'] = df4['station_id'].astype(str).str.lstrip('0')
df3['station_id'] = df3['station_id'].astype(str).str.lstrip('0')

df4['Heavy_traffic'] = df4['station_id'].isin(df3['station_id']).astype(int)

df4

df4.to_csv('dataset/901_traffic.csv', index=True)
