==> 02_NSW_Charger_Optimizer/step2_concat_dataset.ipynb <==
# 这个文件用来生成dataset的坐标和target部分

import pandas as pd
import numpy as np
import geopandas as gpd
from shapely.geometry import Point


df1 = pd.read_csv('dataset/02_NSW_DC_fast_chargers_2023_08-22.csv')
df2 = pd.read_csv('dataset/02_NSW_destination_chargers_2023_08_16.csv',encoding='ISO-8859-1')

df3 = pd.read_csv('dataset/002_neg_point.csv')

df4 = pd.read_csv('dataset/02_approved_Charger_list_v1.csv')
df5 = pd.read_csv('dataset/002_Type5.csv')

boundary_df = gpd.read_file('dataset/boundary.geojson')

###### step1. 整合正类数据，即存在于02_NSW_DC_fast_chargers_2023_08-22.csv(Type1)和02_NSW_destination_chargers_2023_08_16.csv(Type2)的坐标点

df1 = df1[['Latitude', 'Longitude']]
df1['Type'] = 1

df2 = df2[['Latitude', 'Longitude']]
df2['Type'] = 2

result1 = pd.concat([df1, df2], ignore_index=True)

result1

result1[result1['Type']==2]

###### step2. 整合正类数据和负类数据，负类数据存在于05_neg_point.csv(Type3)

result2 = pd.concat([result1, df3], ignore_index=True)

result2

###### step3. 加上需要预测的坐标点，Type记为4(Approved)，5(自己找的)

df_temp = df4[['lat', 'lng']]
df_temp['Type'] = 4
df_temp = df_temp.rename(columns={'lat': 'Latitude', 'lng': 'Longitude'})

df_temp

result3 = pd.concat([result2, df_temp], ignore_index=True)
result3

result3 = pd.concat([result3,df5], ignore_index=True)

###### step3. 对坐标点进行筛选，过滤掉不属于NSW州的坐标点

# 转换DataFrame到GeoDataFrame
gdf = gpd.GeoDataFrame(result3, geometry=gpd.points_from_xy(result3.Longitude, result3.Latitude))

# 使用within方法来检查点是否在NSW边界内
gdf['is_in_nsw'] = gdf.apply(lambda row: boundary_df.contains(row.geometry).any(), axis=1)


gdf

gdf2= gdf[gdf['is_in_nsw']!=True]
gdf2

# 筛选在NSW州内的点
gdf_in_nsw = gdf[gdf['is_in_nsw']]

# 删除不需要的'is_in_nsw'列
gdf_in_nsw = gdf_in_nsw.drop(columns=['is_in_nsw','geometry'])

# 现在 gdf_in_nsw 只包含在NSW边界内的点
gdf_in_nsw

gdf_in_nsw.to_csv('dataset/902_coordinate.csv', index=False)
