==> 02_NSW_Charger_Optimizer/step3_merged_info.ipynb <==
import re
import pandas as pd
import numpy as np
import geopandas as gpd
from shapely.geometry import Point, LineString
from geopy.distance import great_circle

origin_df = pd.read_csv('dataset/902_coordinate.csv')

df11 = pd.read_csv('dataset/031_metro20210906.csv')
df22 = pd.read_csv('dataset/031_regional20210906.csv')

df1 = df11[df11['Time']=='1/1/2023']
df2 = df22[df11['Time']=='1/1/2023']



df3 = pd.read_csv('dataset/901_traffic.csv')

df4_1 = pd.read_csv('dataset/04_Copy_of_POI_(Fast_food).csv')
df4_2 = pd.read_csv('dataset/04_Copy_of_POI_(Fuel_station).csv')
df4_3 = pd.read_csv('dataset/04_Copy_of_POI_(Tourism).csv',encoding='ISO-8859-1')


# df3 = gpd.read_file('dataset/032_route_elevations_optimised.geojson')

###### step1. 坐标是否处于区域充电最佳位置，距离最近的充电最佳中心点的距离，属于Metro1, Major towns2，Journey enablement3 

df1 = df1[["Latitude", "Longitude","Zone"]]
df1['location_type'] = 1
type_mapping1 = {
    "1 km": 3
}
df1['Zone'] = df1['Zone'].map(type_mapping1)
# df1

df2 = df2[["Latitude", "Longitude","Type","Zone (km)"]]
df2.columns = ["Latitude", "Longitude","Type","Zone"]
type_mapping2 = {
    "Major towns": 2,
    "Journey enablement": 3
}
df2['location_type'] = df2['Type'].map(type_mapping2)
type_mapping3 = {
    "20km": 20,
    "10km": 10
}
df2['Zone'] = df2['Zone'].map(type_mapping3)
df2 = df2.drop(columns=['Type'])
# df2

region_df = pd.concat([df1, df2], ignore_index=True)
region_df

# Haversine公式
def haversine(lat1, lon1, lat2, lon2):
    R = 6371  # 地球半径，单位为千米
    lat1, lon1, lat2, lon2 = map(np.radians, [lat1, lon1, lat2, lon2])
    dlat = lat2 - lat1 
    dlon = lon2 - lon1 
    a = np.sin(dlat/2)**2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon/2)**2
    c = 2 * np.arcsin(np.sqrt(a)) 
    km = R * c
    return km

def calculate_distance(df1, df2, feature_name1,feature_name2,feature_name3,feature_name4):
    counts = []
    temps = []
    dists = []
    zones = []
    for index1, row1 in df1.iterrows():
        count = []
        min_dist = 1000000
        min_temp = 0
        zone_temp = 0
        lat1, lon1 = row1['Latitude'], row1['Longitude']
        for index2, row2 in df2.iterrows():
            lat2, lon2 = row2['Latitude'], row2['Longitude']
#             zone_str = str(row2['Zone']).replace('km','')
#             zone_str = zone_str.replace(' ','')
            zone = int(row2['Zone'])
            
            distance = haversine(lat1, lon1, lat2, lon2)
            if distance < min_dist:
                min_dist = distance
                min_temp = row2['location_type']
                zone_temp = zone
            if distance < zone:      ### 如果满足条件，则记录对应的MVA和
                count.append(row2['location_type'])
        dists.append(min_dist)    
        if len(count)==0:
            counts.append(min_temp)
            temps.append(0)
            zones.append(zone_temp)
        elif 1 in count:
            counts.append(1)
            temps.append(1)
            zones.append(3)
        elif 2 in count:
            counts.append(2)
            temps.append(1)
            zones.append(10)
        elif 3 in count:
            counts.append(3)
            temps.append(1)
            zones.append(20)
    df_temp = df1.copy()
    df_temp[feature_name4] = zones
    df_temp[feature_name1] = counts
    df_temp[feature_name2] = temps
    df_temp[feature_name3] = dists
    return df_temp

temp1 = calculate_distance(origin_df, region_df, 'location_type', 'if_optimal_zone','dist_to_optimal_point','Zone')

temp1['location_type'] = temp1['location_type'].astype('int64')

# temp1[temp1['location_type']==1].head(40)
temp1

###### step2. 获取Zone范围内的交通总量以及是否有重型交通

df3

def calculate_traffic_count(df1, df2, feature_name):
    sums = []
    tags = []
    temp_df = df1.copy()
    for index1, row1 in df1.iterrows():
        total_traffic_count = 0
        heavy_count = 0
        lat1, lon1 = row1['Latitude'], row1['Longitude']
        zone = row1['Zone']
        for index2, row2 in df2.iterrows():
            lat2, lon2 = row2['wgs84_latitude'], row2['wgs84_longitude']
            distance = haversine(lat1, lon1, lat2, lon2)
            if distance < zone:
                total_traffic_count += row2['traffic_count']
                heavy_count += row2['Heavy_traffic']
        sums.append(total_traffic_count)
        if heavy_count!=0:
            tags.append(1)
        else:
            tags.append(0)
    temp_df[feature_name] = sums
    temp_df['Heavy_traffic'] = tags
    return temp_df

temp2 = calculate_traffic_count(temp1,df3,'traffic_count')

temp2[temp2['Heavy_traffic']==0]

###### step3. 获取距离最近的已有的快速充电桩和目的地充电桩

chargers_df = origin_df[(origin_df["Type"]==1)|(origin_df["Type"]==2)]
chargers_df

### 计算每个现有的station与最近的station的距离
def calculate_shortest_distance(df1, df2):
    counts1 = []
    counts2 = []
    temp_df = df1.copy()
    for index1, row1 in df1.iterrows():
        count1 = 10000000
        count2 = 10000000
        lat1, lon1, type1 = row1['Latitude'], row1['Longitude'], row1['Type']
        for index2, row2 in df2.iterrows():
            lat2, lon2, type2 = row2['Latitude'], row2['Longitude'], row2['Type']
            distance = haversine(lat1, lon1, lat2, lon2)
            if distance!=0 and type2 == 1 and distance < count1:
                count1 = distance
            if distance!=0 and type2 == 2 and distance < count2:
                count2 = distance
        counts1.append(count1)
        counts2.append(count2)
    temp_df['Distance to the nearest fast charger'] = counts1
    temp_df['Distance to the nearest destination charger'] = counts2
    return temp_df

temp3 = calculate_shortest_distance(temp2,chargers_df)

temp3

###### step4. 获取兴趣点( Fast_food, Fuel_station, Tourism)

def calculate_POI_count(df1, df2, feature_name):
    counts = []
    temp_df = df1.copy()
    for index1, row1 in df1.iterrows():
        count = 0
        lat1, lon1 = row1['Latitude'], row1['Longitude']
        zone = row1['Zone']
        for index2, row2 in df2.iterrows():
            lat2, lon2 = row2['latitude'], row2['longitude']
            distance = haversine(lat1, lon1, lat2, lon2)
            if distance < zone:
                count += 1
        counts.append(count)
    temp_df[feature_name] = counts
    return temp_df

temp4 = calculate_POI_count(temp3,df4_1,'Fast_food')

temp4

temp5 = calculate_POI_count(temp4,df4_2,'Fuel_station')

temp6 = calculate_POI_count(temp5,df4_3,'Tourism')

temp6

# temp6.to_csv('dataset/003_dataset.csv', index=False)

###### step5. 坐标的z_str为距离最近的route的z_str，坐标zone内的route线(是否是交通枢纽)

geo_df = gpd.read_file('dataset/032_route_elevations_optimised.geojson')

temp6['geometry'] = temp1.apply(lambda row: Point(row['Longitude'], row['Latitude']), axis=1)
temp6

def min_distance(point, lines):
    distances = lines.distance(point)
    sorted_distances = distances.sort_values()
    for index in sorted_distances.index:
        z_str = lines.iloc[index]['Z_STR']
        if pd.notna(z_str):
            return sorted_distances.iloc[0], z_str
    return sorted_distances.iloc[0], None

temp6[['dist', 'z_str']] = temp6['geometry'].progress_apply(lambda point: min_distance(point, geo_df)).apply(pd.Series)


temp7 = temp6.drop(columns=['geometry'])

temp7

temp7.to_csv('dataset/903_dataset.csv', index=False)

# gdf1 = gpd.GeoDataFrame(temp1, geometry=gpd.points_from_xy(temp1.Longitude, temp1.Latitude))
# # gdf1.crs = 'EPSG:4326'  # 设置CRS

# def min_distance(point, gdf2):
#     gdf2['distance'] = gdf2.apply(lambda row: point.distance(row.geometry), axis=1)
#     min_row = gdf2.loc[gdf2['distance'].idxmin()]
#     return pd.Series({'dist': min_row['distance'] / 1000, 'z_str': min_row['Z_STR']})  # 转换为千米

# gdf1[['dist', 'z_str']] = gdf1.apply(lambda row: min_distance(row.geometry, df3), axis=1)
