==> 02_NSW_Charger_Optimizer/step4_multi_classi_model.ipynb <==
## dataset_split

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix

df1 = pd.read_csv('dataset/9003_dataset.csv')

df1

###### 只保留Type!=4的部分，删除"Latitude","Longitude"和"Zone"，"Type"做target

df2 = df1[df1['Type']!=4]
df2 = df2[df2['Type']!=5]
df2 = df2.drop(['Latitude', 'Longitude', 'Zone'], axis=1)

df2

y = df2['Type']
X = df2.drop('Type', axis=1)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

###### Linear Regression

from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, r2_score

model = LinearRegression()

model.fit(X_train, y_train)

y_pred = model.predict(X_test)

mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"MSE：{mse:.4f}")
print(f"R^2 score：{r2:.4f}")


###### KNN

from sklearn.neighbors import KNeighborsClassifier

knn_classifier = KNeighborsClassifier(n_neighbors=20)

knn_classifier.fit(X_train, y_train)

y_pred = knn_classifier.predict(X_test)

accuracy = accuracy_score(y_test, y_pred)

conf_matrix = confusion_matrix(y_test, y_pred)
class_report = classification_report(y_test, y_pred)


print(f"The accuracy of KNN is {accuracy:.4f}")
print()
print(class_report)


###### SVM

from sklearn.svm import SVC

# 创建SVM分类器
# svm_classifier = SVC(kernel='linear', C=1.0)  # 选择线性核和适当的C值
svm_classifier = SVC(kernel='poly', degree=3, C=2)  # 使用多项式核，设置阶数为3，C值为1.0


svm_classifier.fit(X_train, y_train)

y_pred = svm_classifier.predict(X_test)

accuracy = accuracy_score(y_test, y_pred)

conf_matrix = confusion_matrix(y_test, y_pred)
class_report = classification_report(y_test, y_pred)

print(f"accuracy：{accuracy:.4f}")
print(class_report)


y_pred

###### Multinomial Logistic Regression

from sklearn.linear_model import LogisticRegression

model = LogisticRegression(multi_class='multinomial', solver='lbfgs')

model.fit(X_train, y_train)

y_pred = model.predict(X_test)

accuracy = accuracy_score(y_test, y_pred)

conf_matrix = confusion_matrix(y_test, y_pred)
class_report = classification_report(y_test, y_pred)

print(f"accuracy：{accuracy:.4f}")
print(class_report)


###### Random Forest

from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, r2_score

rf_model = RandomForestRegressor(random_state=42)

rf_model.fit(X_train, y_train)

y_pred = rf_model.predict(X_test)

# 计算均方误差（Mean Squared Error）和R^2分数来评估模型性能
mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"MSE：{mse:.4f}")
print(f"R^2 score：{r2:.4f}")


from sklearn.ensemble import RandomForestClassifier


rf_classifier = RandomForestClassifier(random_state=42,n_estimators=200,min_samples_split=5)

rf_classifier.fit(X_train, y_train)

y_pred = rf_classifier.predict(X_test)

accuracy = accuracy_score(y_test, y_pred)

conf_matrix = confusion_matrix(y_test, y_pred)
class_report = classification_report(y_test, y_pred)


print(f"The accuracy of Random Forest is {accuracy:.4f}")
print()
print(class_report)


###### LightGBM

import lightgbm as lgb

lgb_classifier = lgb.LGBMClassifier(random_state=42,num_leaves=35,min_data_in_leaf=21)

lgb_classifier.fit(X_train, y_train)

y_pred = lgb_classifier.predict(X_test)

accuracy = accuracy_score(y_test, y_pred)

conf_matrix = confusion_matrix(y_test, y_pred)
class_report = classification_report(y_test, y_pred)

print(f"The accuracy of LightGBM is {accuracy:.4f}")
print()
print(class_report)

###### Neural Networks

y_train = y_train - 1  # 将标签值减去1，从1、2、3转换为0、1、2
y_test = y_test - 1

import tensorflow as tf
from tensorflow import keras



# 创建神经网络模型
model = keras.Sequential([
    keras.layers.Input(shape=(X_train.shape[1],)),  # 输入层
    keras.layers.Dense(64, activation='relu'),  # 隐藏层1
    keras.layers.Dense(32, activation='relu'),  # 隐藏层2
    keras.layers.Dense(3, activation='softmax')  # 输出层，设置单元数量为3，使用softmax激活函数
])

# 编译模型并指定学习率
optimizer = keras.optimizers.Adam(learning_rate=0.01)  # 指定学习率
model.compile(optimizer=optimizer, loss='sparse_categorical_crossentropy', metrics=['accuracy'])

# 在训练集上训练模型
model.fit(X_train, y_train, epochs=20, batch_size=32, validation_split=0.2)

# 在测试集上进行预测
y_pred_prob = model.predict(X_test)
y_pred = tf.argmax(y_pred_prob, axis=1)  # 从概率中选择具有最高概率的类别

accuracy = accuracy_score(y_test, y_pred)

conf_matrix = confusion_matrix(y_test, y_pred)
class_report = classification_report(y_test, y_pred)

print(f"The accuracy of Neural Networks is {accuracy:.4f}")
print()
print(class_report)

###### dataset中Type=4的数据来自02_approved_Charger_list_v1.csv，全为正类，可以输出预测类别为负类占的百分比当作evaluation的指标

df3 = df1[df1["Type"]==4]

df3 = df3.drop(['Latitude', 'Longitude', 'Zone','Type'], axis=1)
df3

predictions = rf_classifier.predict(df3)
probabilities = rf_classifier.predict_proba(df3)

predictions

np.count_nonzero(predictions==3)/len(predictions)

1-np.count_nonzero(predictions==3)/len(predictions)

probabilities

temp = []
for i in probabilities:
    temp.append(max(i))
sum(temp)/len(temp)

###### dataset中Type=5的数据来自002_Type5.csv，为自己手动筛选的可能作为charger的坐标，输出top10作为最优预测

df4 = df1[(df1['Type'] == 4) | (df1['Type'] == 5)]

df5 = df4.drop(['Latitude', 'Longitude', 'Zone','Type'], axis=1)
df5

predictions = rf_classifier.predict(df5)
probabilities = rf_classifier.predict_proba(df5)

predictions

# 找到每行的最大值
# max_values = np.max(probabilities, axis=1)
# 找出最大值大于0.9的行的索引
# indexes = np.where(max_values >= 0.9)[0]
indexes = np.where(np.max(probabilities[:, :2], axis=1) > 0.8)[0]


df_final = df4.iloc[indexes]
df_final

df_final['predictions'] = predictions[indexes]
df_final

df_final[df_final['predictions']==1]

feature_name = X.columns.tolist()
feature_name

importances = rf_classifier.feature_importances_

# 将特征重要性与特征名称结合
feature_importances = zip(feature_name, importances)

# 对特征进行排序
sorted_feature_importances = sorted(feature_importances, key=lambda x: x[1], reverse=True)


for feature, importance in sorted_feature_importances:
    print(f"Feature: {feature}, Importance: {importance:.4f}")

lgb_classifier

importances2 = lgb_classifier.feature_importances_
# feature_importances2 = zip(feature_name, importances2)
# sorted_feature_importances2 = sorted(feature_importances2, key=lambda x: x[1], reverse=True)
# 归一化特征重要性
total_importance = sum(importances2)
normalized_importances2 = importances2 / total_importance

feature_importances2 = zip(feature_name, normalized_importances2)
sorted_feature_importances2 = sorted(feature_importances2, key=lambda x: x[1], reverse=True)

for feature, importance in sorted_feature_importances2:
    print(f"Feature: {feature}, Importance: {importance:.4f}")

for feature, importance in sorted_feature_importances2:
    print(f"Feature: {feature}, Importance: {(importance)}")
